Add files via upload
cmd_gen.m is a  GNU Octave script to cross-identify stars in two different photometry catalogues by positional coincidence. It produces a color-magnitude diagram. The data sets include 130,000 by 20 arrays of floating point numbers. A double for loop over the RAs and Decs of the objects in each list would have required 16.9 billion comparisons to make the identifications. I vectorized the loops, speeding up the comparison by a factor several thousand. That is, I carried out the comparison by element-wise operations on an array containing n copies (horizontally) of an m-element vector of magnitudes in a particular filter and an array containing m copies (vertically) of an n-element vector of magnitudes in another filter. I then selected elements from the resulting array that matched the required distance tolerance.