color_dist_p.m
A script that produces color histograms from color magnitude arrays and fits a Gaussian mixture model to them, using an EM algorithm. This program includes a subprogram that plots the color-magnitude diagram onto which the user draws a ridge line which is used to verticalize the color-magnitude diagram prior to making the histogram.