fit_ridge_line_p.m
This script allows points to be marked on a color magnitude diagram that mark a locus of maximum point densities (the ridge-line). It then smoothly connects them with a 3rd or 4th order spline. The spline-fit parameters are then used in the script color_dist_p.m to interpolate magnitudes corresponding to fiducial colors along the ridge-line. These are used to subtract the fiducial colors at the magnitudes of the stars from the colors of those stars, to verticalize the color-magnitude diagram.