geotran_p.m
A script that compares two star catalogues to cross identify stars by overlaying maps of the fields and, and pattern-matching, to find the geometric transformation required to line up one catalog with the other.  This program makes use of weighted non-linear least-squares fitting to estimate the difference in their pointings (i.e. del_RA, del_dec between directions of optical centers) rotation angle, and magnification required to overly one field on the other and the magnitude offset between the catalogues.